requirements.txt
django
djangorestframework
postgres
drf-yasg
